move serialization helpers to their own file
