+++ include/json.hpp
@@ -6,10 +6,6 @@
 #include <vector>
 #include <string>
 #include <stdexcept>
-#include <unordered_set>
-#include <set>
-#include <map>
-#include <unordered_map>
 
 namespace json {
 	enum class Type {
@@ -218,151 +214,4 @@ namespace json {
 	inline Value parse(std::string_view source) {
 		return Value::from_str(source);
 	}
-
-	// allow converting parsing JSON directly to STL containers for convenience
-
-	template <class T>
-	struct Serialize<std::vector<T>> {
-		static std::vector<T> from_json(Value const& value)
-			requires requires(Value const& value) {
-				value.template as<std::decay_t<T>>();
-			}
-		{
-			std::vector<T> res;
-			auto arr = value.as_array();
-			res.reserve(arr.size());
-			for (auto& item : arr) {
-				res.push_back(item.template as<std::decay_t<T>>());
-			}
-			return res;
-		}
-
-		static Value to_json(std::vector<T> const& value)
-			requires requires(T const& value) {
-				Value(value);
-			}
-		{
-			Array res;
-			res.reserve(value.size());
-			for (auto& item : value) {
-				res.push_back(Value(item));
-			}
-			return res;
-		}
-	};
-
-	template <class T>
-	struct Serialize<std::unordered_set<T>> {
-		static std::unordered_set<T> from_json(Value const& value)
-			requires requires(Value const& value) {
-				value.template as<std::decay_t<T>>();
-			}
-		{
-			std::unordered_set<T> res;
-			auto arr = value.as_array();
-			res.reserve(arr.size());
-			for (auto& item : arr) {
-				res.insert(item.template as<std::decay_t<T>>());
-			}
-			return res;
-		}
-
-		static Value to_json(std::unordered_set<T> const& value)
-			requires requires(T const& value) {
-				Value(value);
-			}
-		{
-			Array res;
-			res.reserve(value.size());
-			for (auto& item : value) {
-				res.push_back(Value(item));
-			}
-			return res;
-		}
-	};
-
-	template <class T>
-	struct Serialize<std::set<T>> {
-		static std::set<T> from_json(Value const& value)
-			requires requires(Value const& value) {
-				value.template as<std::decay_t<T>>();
-			}
-		{
-			std::set<T> res;
-			auto arr = value.as_array();
-			for (auto& item : arr) {
-				res.insert(item.template as<std::decay_t<T>>());
-			}
-			return res;
-		}
-
-		static Value to_json(std::set<T> const& value)
-			requires requires(T const& value) {
-				Value(value);
-			}
-		{
-			Array res;
-			res.reserve(value.size());
-			for (auto& item : value) {
-				res.push_back(Value(item));
-			}
-			return res;
-		}
-	};
-
-	template <class T>
-	struct Serialize<std::map<std::string, T>> {
-		static std::map<std::string, T> from_json(Value const& value)
-			requires requires(Value const& value) {
-				value.template as<std::decay_t<T>>();
-			}
-		{
-			std::map<std::string, T> res;
-			auto obj = value.as_object();
-			for (auto& [k, v] : obj) {
-				res.insert({ k, v.template as<std::decay_t<T>>() });
-			}
-			return res;
-		}
-
-		static Value to_json(std::map<std::string, T> const& value)
-			requires requires(T const& value) {
-				Value(value);
-			}
-		{
-			Object res;
-			for (auto& [k, v] : value) {
-				res[k] = (Value(v));
-			}
-			return res;
-		}
-	};
-
-	template <class T>
-	struct Serialize<std::unordered_map<std::string, T>> {
-		static std::unordered_map<std::string, T> from_json(Value const& value)
-			requires requires(Value const& value) {
-				value.template as<std::decay_t<T>>();
-			}
-		{
-			std::unordered_map<std::string, T> res;
-			auto obj = value.as_object();
-			for (auto& [k, v] : obj) {
-				res.insert({ k, v.template as<std::decay_t<T>>() });
-			}
-			return res;
-		}
-
-		static Value to_json(std::unordered_map<std::string, T> const& value)
-			requires requires(T const& value) {
-				Value(value);
-			}
-		{
-			Object res;
-			for (auto& [k, v] : value) {
-				res[k] = (Value(v));
-			}
-			return res;
-		}
-	};
 }
